aider: Fixed YAML syntax issue in `vm-setup/roles/pip_pkgs/tasks/main.yml`.
# Aider chat conversation:

USER: I ran this command:

 cd vm-setup && ansible-playbook -K sweetums.yml

And got this output:

ERROR! We were unable to read either as JSON nor YAML, these are the errors we got from each:
JSON: Expecting value: line 1 column 1 (char 0)

Syntax Error while loading YAML.
  did not find expected key

The error appears to be in '/home/harleypig/projects/dotfiles/vm-setup/roles/pip_pkgs/tasks/main.yml': line 10, column 4, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

   - name: Install python3-pip
   ^ here

ASSISTANT:
^C KeyboardInterrupt

USER: Try again

ASSISTANT: The error message indicates that there is a YAML syntax issue in the `vm-setup/roles/pip_pkgs/tasks/main.yml` file. The problem is likely due to incorrect indentation or a misplaced dash (`-`) in the task list. YAML is sensitive to indentation, and tasks in an Ansible playbook should be properly indented as list items.

Here is the corrected diff for the `vm-setup/roles/pip_pkgs/tasks/main.yml` file:

```diff
--- vm-setup/roles/pip_pkgs/tasks/main.yml
+++ vm-setup/roles/pip_pkgs/tasks/main.yml
@@ -7,7 +7,7 @@
    vars:
      all_groups: "{{ pip_package_groups.keys() }}"
      exclude_groups: "{{ exclude_group | default([]) }}"
      is_exclusion_list: "{{ is_exclude_list | default(false) }}"
      requested_groups: "{{ all_groups if not is_exclusion_list else all_groups | difference(exclude_groups) }}"

-   - name: Install python3-pip
+  - name: Install python3-pip
      apt:
        name: python3-pip
        state: present

    block:
```

This change fixes the indentation of the task `- name: Install python3-pip` to align correctly with the other tasks in the list.