aider: Modify the gnome settings role to expect the location of the Tilix profile and add the setting to the sweetums and work playbooks.
# Aider chat conversation:

USER: Modify the gnome settings role so it is expecting the location of the tilix profile to be used and add that setting to the sweetums and work playbooks.

ASSISTANT: First, we will modify the `gnome_settings` role to expect the location of the Tilix profile to be passed as a variable. We will add a default value for the variable in the `defaults/main.yml` file of the role and then use that variable in the `tilix_profile.yml` task file.

Here are the diffs for the necessary changes:

For the `gnome_settings` role's `defaults/main.yml`:

```diff
--- vm-setup/roles/gnome_settings/defaults/main.yml
+++ vm-setup/roles/gnome_settings/defaults/main.yml
@@ ... @@
 gnome_settings_show_trash: false
 gnome_settings_show_home: false
+tilix_profile_filename: "{{ playbook_dir }}/tilix-profile.dconf"
 gnome_settings_desktop_background:
   - picture-uri ''
   - picture-uri-dark ''
```

And for the `gnome_settings` role's `tasks/tilix_profile.yml`:

```diff
--- vm-setup/roles/gnome_settings/tasks/tilix_profile.yml
+++ vm-setup/roles/gnome_settings/tasks/tilix_profile.yml
@@ ... @@
 - name: Check if Tilix profile file exists
   stat:
-    path: "{{ tilix_profile_filename }}"
+    path: "{{ gnome_settings_tilix_profile_filename }}"
   register: tilix_profile_file

 - name: Load Tilix profile if file exists
   shell: "dconf load /com/gexperts/Tilix/ < {{ gnome_settings_tilix_profile_filename }}"
   when: tilix_profile_file.stat.exists
   notify: Restart Tilix
```

Next, we will add the setting for the Tilix profile location to both the `sweetums.yml` and `work.yml` playbooks.

For the `sweetums.yml` playbook:

```diff
--- vm-setup/sweetums.yml
+++ vm-setup/sweetums.yml
@@ ... @@
   roles:
     - role: install_pkgs
     - role: pip_pkgs
     - role: nerd_font
     - role: add_iac
     - role: git_bug
-    - role: gnome_settings
+    - role: gnome_settings
+      vars:
+        gnome_settings_tilix_profile_filename: "{{ playbook_dir }}/tilix-profile.dconf"
```

And for the `work.yml` playbook:

```diff
--- vm-setup/work.yml
+++ vm-setup/work.yml
@@ ... @@
   roles:
     - role: install_pkgs
       is_exclude_list: true
       exclude_group:
         - hledger
     - role: pip_pkgs
       is_exclude_list: true
       exclude_group:
         - linode
     - role: nerd_font
     - role: add_iac
     - role: git_bug
-    - role: gnome_settings
+    - role: gnome_settings
+      vars:
+        gnome_settings_tilix_profile_filename: "{{ playbook_dir }}/tilix-profile.dconf"
```

These changes will allow the `gnome_settings` role to use the `tilix_profile_filename` variable that is set in the playbooks, which can be overridden as needed for different environments or use cases.