aider: Fixed YAML syntax error in `vm-setup/roles/nerd_font/tasks/main.yml`.
# Aider chat conversation:

USER: I ran this command:

 cd vm-setup && ansible-playbook -K nerdfont.yml

And got this output:

ERROR! We were unable to read either as JSON nor YAML, these are the errors we got from each:
JSON: Expecting value: line 1 column 1 (char 0)

Syntax Error while loading YAML.
  did not find expected '-' indicator

The error appears to be in '/home/harleypig/projects/dotfiles/vm-setup/roles/nerd_font/tasks/main.yml': line 21, column 3, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

    path: "{{ nerd_font_local_filename }}"
  register: file_check
  ^ here

ASSISTANT: The error message indicates that there is a syntax issue in the `vm-setup/roles/nerd_font/tasks/main.yml` file, specifically around line 21. It seems that the YAML syntax is incorrect due to improper indentation or a missing `-` indicator for a list item. I will provide a diff to correct the syntax error in the file.

```diff
--- vm-setup/roles/nerd_font/tasks/main.yml
+++ vm-setup/roles/nerd_font/tasks/main.yml
@@ ... @@
 - name: Get the latest release version of Nerd Font
   uri:
     url: "{{ nerd_font_version_url }}"
     method: GET
     return_content: yes
     status_code: 200
     headers:
       Accept: "application/vnd.github.v3+json"
   register: nerd_font_latest_release
   changed_when: false

 - name: Set Nerd Font download URL and version
   set_fact:
     nerd_font_version: "{{ nerd_font_latest_release.json.tag_name }}"
     nerd_font_url: "https://github.com/ryanoasis/nerd-fonts/releases/download/{{ nerd_font_version }}/{{ nerd_font_name }}.zip"
   when: nerd_font_latest_release is changed

--  stat:
-    path: "{{ nerd_font_local_filename }}"
-  register: file_check
-  changed_when: false
+- name: Check if Nerd Font already installed
+  stat:
+    path: "{{ nerd_font_local_filename }}"
+  register: file_check
+  changed_when: false
+
```

This diff corrects the indentation and adds the missing `- name:` key to the task, which should resolve the syntax error you encountered. Apply this patch to the `vm-setup/roles/nerd_font/tasks/main.yml` file and try running the playbook again.