aider: Add task to handle GitHub API rate limit exceeded in getting the latest Nerd Font version.
# Aider chat conversation:

USER: Review the code for the task to get the latest nerd font version. Are there any problems with that?

ASSISTANT: Upon reviewing the task to get the latest Nerd Font version in the `vm-setup/roles/nerd_font/tasks/main.yml` file, I noticed that the task is well-defined and should work as expected. It uses the `uri` module to make a GET request to the GitHub API for the latest release of Nerd Fonts, and it registers the result in the `nerd_font_latest_release` variable.

However, there is a potential improvement that can be made. The task does not currently handle the situation where the GitHub API rate limit is exceeded. When the rate limit is exceeded, the API returns a 403 status code with a message indicating that the rate limit has been hit. It would be beneficial to add a task that checks for this condition and provides a more informative error message.

Here is the proposed change to handle the rate limit:

```diff
--- vm-setup/roles/nerd_font/tasks/main.yml
+++ vm-setup/roles/nerd_font/tasks/main.yml
@@ ... @@
 - name: Get the latest release version of Nerd Font
   uri:
     url: "{{ nerd_font_version_url }}"
     method: GET
     return_content: yes
     status_code: 200
     headers:
       Accept: "application/vnd.github.v3+json"
   register: nerd_font_latest_release
   changed_when: false
+
+- name: Fail if GitHub API rate limit is exceeded
+  fail:
+    msg: "GitHub API rate limit exceeded. Please try again later."
+  when: nerd_font_latest_release.status == 403
```

This additional task will check the status code of the response from the GitHub API, and if it is 403, it will fail the play with a message indicating that the rate limit has been exceeded. This will make it clearer to the user why the task failed if they encounter this issue.