aider: Modified tasks in Nerd Font role to run only if `nerd_font_filename` needed to be downloaded.
# Aider chat conversation:

USER: The tasks in the nerd font role from the compare versions tasks onward need to be modified to only run if nerd_font_filename needed to be downloaded.

ASSISTANT: To modify the tasks in the Nerd Font role to only run if `nerd_font_filename` needed to be downloaded, we need to add a condition to each task based on whether the file was downloaded or not. This can be done by checking the `changed` status of the task that downloads the font file.

Here is the diff for the necessary changes:

```diff
--- vm-setup/roles/nerd_font/tasks/main.yml
+++ vm-setup/roles/nerd_font/tasks/main.yml
@@ ... @@
 - name: Compare installed and latest Nerd Font versions
   set_fact:
     nerd_font_version_mismatch: "{{ installed_nerd_font_version != nerd_font_version }}"

 - name: Skip download if the installed version matches the latest version
   meta: end_play
   when: not nerd_font_version_mismatch

+- name: Download Nerd Font if version mismatch
+  get_url:
+    url: "{{ nerd_font_url }}"
+    dest: "{{ nerd_font_dir }}/{{ nerd_font_filename }}"
+    mode: '0644'
+  when: nerd_font_version_mismatch
+  register: nerd_font_download_result
+
 - name: Ensure Nerd Fonts directory exists
   file:
     path: "{{ nerd_font_dir }}"
     state: directory
-  when: nerd_font_version_mismatch
+  when: nerd_font_download_result is changed

 - name: Unarchive Nerd Font if version mismatch
   unarchive:
     src: "{{ nerd_font_url }}"
     dest: "{{ nerd_font_dir }}"
     remote_src: yes
-  when: nerd_font_version_mismatch
+  when: nerd_font_download_result is changed

 - name: Update font cache
   command: fc-cache -fv
-  when: nerd_font_version_mismatch
+  when: nerd_font_download_result is changed
```

Apply this patch to the `vm-setup/roles/nerd_font/tasks/main.yml` file. This will ensure that the tasks for ensuring the Nerd Fonts directory exists, unarchiving the Nerd Font, and updating the font cache will only run if the Nerd Font file was actually downloaded.